#1 refactor code and align to conventions
