+++ applications/areg-portal/frontend/src/services/data.json
