--- applications/areg-portal/frontend/src/utils/constants.ts
@@ -1,1 +1,0 @@
-
